weather_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
pydantic==2.5.0
sqlalchemy==2.0.27
asyncpg==0.29.0
python-dotenv==1.0.0
email-validator==2.1.0